+++ src/newtab/actions/grid.test.ts
@@ -1,35 +1,62 @@
-import configureStore from 'redux-mock-store'
+import configureStore, { MockStoreEnhanced } from 'redux-mock-store'
 import * as types from '../constants/types'
 import * as actions from './grid'
 
 const mockStore = configureStore()
+let store: MockStoreEnhanced<any, any>
 
 describe('store', () => {
-  it('should dispatch addTile correctly', () => {
+  beforeEach(() => {
     const initialState = {}
-    const store = mockStore(initialState)
+    store = mockStore(initialState)
+  })
 
+  afterEach(() => {
+    store.clearActions()
+  })
+
+  it('should dispatch addTile correctly', () => {
     store.dispatch({ type: types.ADD_TILE })
 
     const storeActions = store.getActions()
-    const expectedPayload = { addTile: actions.addTile }
-    expect(actions).toEqual(expectedPayload)
-  })
+    const expectedPayload = { type: types.ADD_TILE }
+    expect(storeActions).toEqual([expectedPayload])
+  }),
+    it('should dispatch addColumn correctly', () => {
+      store.dispatch({ type: types.ADD_COLUMN })
+
+      const storeActions = store.getActions()
+      const expectedPayload = { type: types.ADD_COLUMN }
+      expect(storeActions).toEqual([expectedPayload])
+    })
 })
 
 describe('actions', () => {
   it('should create an action to add a tile', () => {
-    const url = 'link.url'
     const id = 'tile.id'
+    const url = 'tile.url'
     const expectedAction = {
       type: types.ADD_TILE,
       payload: {
         tile: {
-          url,
           id,
+          url,
         },
       },
     }
-    expect(actions.addTile(url, id)).toEqual(expectedAction)
-  })
+    expect(actions.addTile(id, url)).toEqual(expectedAction)
+  }),
+    it('should create an action to add a column', () => {
+      const id = 'column.id'
+      const expectedAction = {
+        type: types.ADD_COLUMN,
+        payload: {
+          column: {
+            id,
+            tileIds: [],
+          },
+        },
+      }
+      expect(actions.addColumn(id)).toEqual(expectedAction)
+    })
 })