+++ src/newtab/actions/grid.ts
@@ -1,7 +1,12 @@
-import { ADD_TILE } from '../constants/types'
-import { ITile } from '../models/newtab'
+import { ADD_COLUMN, ADD_TILE } from '../constants/types'
+import { IColumn, ITile } from '../models/newtab'
 
 // Define the Action interfaces representing the return types of the action creators
+export interface IAddColumnAction {
+  type: 'ADD_COLUMN'
+  payload: { column: IColumn }
+}
+
 export interface IAddTileAction {
   type: 'ADD_TILE'
   payload: { tile: ITile }
@@ -10,11 +15,27 @@ export interface IAddTileAction {
 // Define the action creators
 
 /**
- * Add a tile to the grid
- * @param url The url link the tile should point to
- * @param id The id of the tile, constructed from a hash of the url and the current time
+ * Add a column to the grid
+ * @param id The id of the column
  */
-export function addTile(url: string, id: string): IAddTileAction {
+export function addColumn(id: string): IAddColumnAction {
+  return {
+    type: ADD_COLUMN,
+    payload: {
+      column: {
+        id,
+        tileIds: [],
+      },
+    },
+  }
+}
+
+/**
+ * Add a tile to the grid
+ * @param url The url the tile should point to
+ * @param id The id of the tile
+ */
+export function addTile(id: string, url: string): IAddTileAction {
   return {
     type: ADD_TILE,
     payload: {
@@ -27,4 +48,4 @@ export function addTile(url: string, id: string): IAddTileAction {
 }
 
 // Define action type for use in reducer
-export type Action = IAddTileAction
+export type Action = IAddColumnAction | IAddTileAction