+++ src/newtab/constants/types.ts
@@ -1,2 +1,3 @@
 // ActionTypes for grid state
+export const ADD_COLUMN = 'ADD_COLUMN'
 export const ADD_TILE = 'ADD_TILE'